Fix errors in integration test runner script.
There are a couple of small errors in the integration testing script
causing the tests to fail.

The first is a mis-placed glob inside of a double quoted string; instead
of copying all of the files in a directory to a tempdir the test runner
tried to copy the literal file "test/integration/tmp/*", and would
subsequently fail to run terraform since no configurations were present.
This has been corrected by moving the asterisk outside of the quotes so
that it can be properly expanded.

The second issue is an omitted curly brace in the templated
`outputs.tf`, which has been restored.