Remove default service account
The ability to manage multiple service accounts with the new
service_accounts submodule means that the explicit resources managing
the groups and roles for the `project-service-account` are no longer
necessary. This commit removes the unneeded resources in favor of
defining the default service account in terms of a default value for the
`service_accounts` variable.