Parametrize inspec tests with attributes
The first pass of inspec test coverage for the project factory used
environment variables to transport information from the test
configuration to inspec, but this was a brittle arrangement since the
use of environment variables were scattered across the control files.
This commit refactors the testing to generate an inspec attributes YAML
file so that test inputs are explicit. This parameterization also opens
up the possibility of letting users reuse these same tests for
validating their usage of the project factory within their environment.