Require `domain` variable for test fixture module
The BATS based integration tests only required a gsuite admin email
address and could extract the GAE auth domain from that address, but
converting tests to test-kitchen removed the ability to easily
extract strings and generate additional variables. While we could do
this work inside of the terraform code, it's simpler to just add
`domain` as a required variable and avoid the hassle entirely.