Update InSpec variables for pinned kitchen-terraform version
kitchen-terraform ran into some [incompatibilities][kt-272] with InSpec
and in response pinned the version of InSpec to 2.2.78; however the
first pass of InSpec tests were written using terraform 2.2.101+, and
depended on features that are not available in the pinned version. This
commit amends some of the InSpec attribute declarations to conform to
the older syntax.

[kt-272]: https://github.com/newcontext-oss/kitchen-terraform/issues/272