Don't check environment when setting InSpec attributes
In order to test the full set of behaviors within the project factory,
we need to test multiple independent Terraform configurations. This is
being implemented through a single set of environment variables that the
different configurations selectively use or ignore. However, this means
that some environment variables may be set, but were intentionally
ignored by the minimal module.

This becomes a problem when InSpec attributes were defined that pulled
attribute default values from the environment when the InSpec attributes
were not set. However these attributes were specifically unset, so using
the environment defined values causes tests to run for features that
were explicitly not created - so we get test failures.

This commit fixes the issue by always setting a default value to `nil`
instead of the environment. This has the additional benefit of reducing
the number of sources for InSpec attributes.