Test project-factory with multiple configurations
Bugs such as #18 are only triggered when the project-factory is used
with defaulted variables, and only before the GCP resources are created.
In order to verify the absence of bugs stemming from unset variables we
have to stand up a separate Terraform configuration.

This commit restructures the test-kitchen tests to use two platforms - a
`full` platform that will enable as many features as possible, and a
`minimal` platform that handles a minimal number of variables. Both
platforms source their configuration from Terraform variables, so
running two platforms requires no additional configuration.