Add troubleshooting guide
The Project Factory interacts with a number of APIs and requires fairly
broad permissions and a number of enabled APIs in order to operate. If
some of these permissions or services or absent the Project Factory can
fail to run, or partially generate a project and then get stuck.

This commit adds documentation for some of the more common errors and
misconfigurations, explains the cause of the errors, and indicates how
to fix them.