Convert test-kitchen to use multiple suites
The [terraform-google-container-vm][container-vm] project implemented a
similar pattern for testing multiple configurations as was implemented
here, but it implemented the different scenarios as different suites
rather than different platforms. This behavior is more consistent with
the behavior that we're testing, in that we're interested in running two
suites, but we only have one platform.

This commit converts the project-factory test-kitchen configuration to
match the container-vm approach and use multiple suites instead of
multiple platforms. Since the tests were already implemented so that
they could be used with different configurations, we symlink
`test/integration/full` and `test/integration/minimal` to
`test/integration/default` to reduce the footprint of this change.

[container-vm]: https://github.com/terraform-google-modules/terraform-google-container-vm/pull/2