Refactor and rename helpers/version-repo.sh
Script is now "release-new-version.sh". Its scope is extended beyond
updating the README version and creating a tag locally. It now commits
the README change(s), pushes that commit, and pushes the tag to the
origin remote - which is assumed to be Github.

The act of pushing a new tag to Github is how new module versions are
released. Tag creation envokes a webhook that will make the new release
appear on registry.terraform.io in about a minute.