Update getting started documentation
Working through the getting started information, I encountered the
following two friction points:

1. I wasn't sure if HOST_PROJECT in the setup-sa.sh script referred to the same
   thing as the Seed Project mentioned in the documentation.  It does.
2. I didn’t understand when the billing account ID optional argument of the
   setup-sa script is required, or not.  Upon investigation the billing account
   ID is required when the Seed Service Account is in a different organization
   from the billing account.

This patch addresses these two points by clarifying terminology and unifying
terms into a Glossary for the whole project.  HOST_PROJECT is removed in favor
of Seed Project.  The troubleshooting guide is updated to clarify the situations
where the billing account argument is required, and the consequences of
proceeding without it.