Fix counts which can not be computed
`gsuite_group` does not need to be included in `core_project_factory`
since the only relevant output is `email`.

The `manage_group` variable on `core_project_factory` removes the need
to check the value of the `group_name` variable which was causing
an issue during graph resolution.

The `count` of the `additive_service_account_grant_to_group` in the
`full` fixture is removed because the `group_email` output will always
be defined in that fixture.