Fix two race conditions
When a new G Suite group is created to manage a project, `core_project_factory`
would attempt to assign the group IAM permissions before the group was
finished being created by the `gsuite_enabled` module. To fix this condition,
an implicit dependency was added to the Terraform using the email attribute
from the `gsuite_group` resource.

Also, the `google_compute_default_service_account` resource depends on the
Compute Engine API being enabled so it is possible for the fetch of the data
resource to fail because it attempts to query the Compute Engine API before it
is fully enabled. Adding an explicit dependency on the services being enabled
fixes this issue.