Move CI configuration logic to script
This commit extracts the CI setup logic into a shell script in this
repository, and updates the Make targets to pass around the correct
environment variables. This change allows us to more closely simulate
how CI works, and allows us to iterate on the CI workflow in pull
requests rather than having to manage everything within the pipeline
definition.