Test for shared VPC subnet bindings
Previously the full suite of tests had tests to ensure that the
roles/compute.networkUser role wasn't bound at the shared VPC host
project, but it did not include a test to ensure that the role was
bound at the subnetwork level. This commit adds a test to check for the
subnetwork level bindings in the event that `var.shared_vpc` and
`var.shared_vpc_subnets` are both passed to project factory.