Refactor tests for randomized names
Without this commit the integration test fixtures will create VPC
networks with randomized names, but the subnetwork names will remain
static. Because subnetworks are regional resources, you can't have two
subnetworks with the same name tied to the same project and region. This
means that without adding randomness to subnetwork names, only 1 run of
the integration test pipeline can occur at any given time.

Normally a random_string resource is used to impart randomness into
resource names, but the way the Network module performs lookups on the
`var.subnets` and `var.secondary_ranges` input variables, those
variables cannot contain computed values. To remedy this, the test
fixtures now include an input variable named
`var.random_string_for_testing` that accepts a string to be appended on
both the network and subnetwork names. The `test/ci_integration.sh`
script has been modified to look for an environment variable named
`RANDOM_STRING_FOR_TESTING` and pass that to
`var.random_string_for_testing`. If that environment variable has not been
declared then a 5-character string is generated on the fly and used.

The random_string resource has been deleted completely in favor of
`var.random_string_for_testing` so that all resources generated in
an integration test run can use the same random characters (to
make it easier to track down all the resources from a single run
if we ever need to do that in the future).