Use project_services module and convert to for_each
Convert the project_services module to use for_each instead of count,
and use that module in the core project factory.

Remove the var.apis_authority option.

Add new helpers/migrate4.py script to convert this project services
migration. To make this migration work, it needs to read values from a
resource, which means calling terraform state show, which is slow.

Doing the migrate with a terraform mv command is blocked by
hashicorp/terraform#22301, so the script instead runs state rm and then
import.
Note that after the migrate each resource instance will show
disable_dependent_services and disable_on_destroy need to be set to
true, because as a Terraform state only value, it can't be gotten from
importing. This makes this workaround functional but messy.

This is also partially blocked by hashicorp/terraform#10462. As this
moved the project services into a module, the depends_on in the
project_service resource block can't be used in the module block because
of this issue. Since the project_id parameter is referencing the
google_project, it might still work correctly.

Resolves #282