Use default statefile and fix dependency breakage
With all the features of Terraform 0.12, the migrate script can
completely run off of remote state, so by default the script will let
Terraform use its default state, remote or local. There is an option to
override this for testing or for keeping remote state safe.

Add verbose option to migrate script, which will print the commands
being run even if dryrun is not enabled.

Revert the output var change in product_services, since it's needed for
the dependency graph. However, the old '.*.' operator doesn't work with
for_each, so replace it with a new [for] construct.