implement prconditions.sh such that it doesn't trip lint failures
correct usage wouldn't wrap an array in quotes when each value is intended
to be added separately, but the linter can't be convinced of that. So
I added quotes but then make sure not to include the array at all if it is
empty in order to avoid adding '' to the argument list.