Fix module crash on empty api list
If the activate_apis variable was set to an empty list, the
project_services module would crash on trying to get the 0 index of the
empty list. Fix is to instead of only returning the project id from the
input if the enable_apis is false, just concat that variable to the
list, so if there is nothing else in the list, that is returned.

Fixes #300