Add tests for disabling default compute SA
The new disable compute service account feature needs tests, so that
option is picked for the minimal suite, and the old delete option is
picked for the full suite. It doesn't look like the full test suite is
run though?

Update the version of the test image to have the newest version of
gcloud, since disabling service accounts only just went GA.