Change default for default compute SA to disable
This is a better option than "delete", since it has the same outcome
(the account doesn't work), but it is possible to get the account back,
unlike which delete which can be unrecoverable. This then should be the
default, since it achieves the same outcome, but is completely and
easily reversible for those that realize they don't want the default
later on.