Simplify project_id output in core_project_factory
I ran into a case where somehow `google_compute_shared_vpc_service_project.shared_vpc_attachment` was an "empty tuple" in the state, and this output was failing because it was trying to index into it. I came across https://github.com/hashicorp/terraform/issues/25578, which might be related, but it does still happen with the latest version of Terraform as of opening this bug (v0.14.9).

Regardless, it seems to me that all this is doing is building a list then taking the first element, so the one-liner should be equivalent.