Process roles for the compute.googleapis.com service identity separately from other service identities when passed via activate_api_identities.
This service identity cannot be created with the `google_project_service_identity` resource. Attempting to activate this service identity leads to an error response. It seems that this API call is not permitted despite lack of documentation. This can be seen by calling:

```sh
gcloud alpha services identity create --service=compute.googleapis.com --log-http
```

Calling this same command with, e.g., cloudbuild.googleapis.com does not result in an error.

It seems the service is activated by default, so instead just retrieve the email using `google_compute_default_service_account`. So get this email, create a combined map with this service identity (if present) and other service identities (if present), and then process this map as before, adding each role for each service identity.